fix(Asset) resolve from group path
While the `Asset::add_to_group_path()` method _should_ be used while the
Asset is being registered, nothing currently prevents (e.g. an
exception) the method from being called on the `Asset` object after
registration.

Since the path resolution in the `get_root_path` method is done
lazily, when the method is called, the fact the Asset has been added to
a specific group path should matter.

This updates the code to support correctly, in my opinion, calls to the
`Asset::add_to_group_path` method after the object registration and take
the group path into account when resolving the root path.

See the test in this PR to have a better understanding.