Prevent assets from being added to assets array twice
When adding assets, an asset would be added to the `Assets::$assets` array with the slug as the index and with a numeric index. There's no reason for the numeric index, as it just causes extra looping and attempts to enqueue.

WP must be outputting script tags with single quotes now. Updated tests to reflect that.